requirements.txt
pandas
openpyxl
python-dotenv
numpy
torch
vllm
tqdm
requests
nltk
python-Levenshtein